Visual Cryptography for Text/requirements.txt
numpy==1.24.0
Pillow==9.3.0
streamlit==1.16.0

